+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## master (unreleased)
+
+* Support symbols passed as metadata, e.g:
+
+        describe 'foo', :focus do
+        end
+
 ## 0.0.1
 
 * Initial release.